+++ core/modules/catalog/helpers/associatedProducts/index.ts
@@ -1,0 +1,7 @@
+import setGroupedProduct from './setGroupedProduct'
+import setBundleProducts from './setBundleProducts'
+
+export {
+  setGroupedProduct,
+  setBundleProducts
+}