+++ core/modules/catalog/helpers/variant/index.ts
@@ -1,0 +1,11 @@
+import omitSelectedVariantFields from './omitSelectedVariantFields'
+import getSelectedVariant from './getSelectedVariant'
+import isOptionAvailable from './isOptionAvailable'
+import findConfigurableVariant from './findConfigurableVariant'
+
+export {
+  omitSelectedVariantFields,
+  getSelectedVariant,
+  isOptionAvailable,
+  findConfigurableVariant
+}