+++ src/themes/default/components/core/ProductPrice.vue
@@ -48,7 +48,7 @@ export default {
     },
     customOptionsPriceDelta () {
       const priceDelta = getCustomOptionPriceDelta(
-        getCustomOptionValues(Object.values(this.customOptions), this.product.custom_options),
+        getCustomOptionValues(Object.values(this.customOptions), this.product.custom_options || []),
         this.product
       )