+++ src/themes/default/components/theme/blocks/Reviews/ReviewsList.vue
@@ -36,7 +36,7 @@
 
 <script>
 import { currentStoreView } from '@vue-storefront/core/lib/multistore'
-import {reviewJsonLd} from '@vue-storefront/core/helpers'
+import { reviewJsonLd } from '@vue-storefront/core/helpers'
 export default {
   props: {
     perPage: {